Dpwgan

+++ download.py
@@ -1,12 +1,26 @@
 import subprocess
-from secrets import SERVER_NAME, REMOTE_LOCAL_PATHS, PROXY, DOWNLOADS_DIR, BKP_DIR
-import os, shutil, time
+from secrets import SERVER_NAME, PROXY, REMOTE_LOCAL_PATHS, DOWNLOADS_DIR, BKP_DIR
+from secrets import SERVER_NAME, PROXY, REMOTE_LOCAL_PATHS_DP, DOWNLOADS_DIR_DP, BKP_DIR_DP
+import os, shutil, time, argparse
 
-REMOTE_ROOT_PATH = SERVER_NAME+':msc-thesis/'
+parser = argparse.ArgumentParser()
+parser.add_argument("--quick", type=lambda v: v=='True', default=True, help="results file only")
+parser.add_argument("--dp", type=lambda v: v=='True', default=False, help="results file only")
+opt = parser.parse_args()
+print(opt)
+
+if opt.dp:
+    REMOTE_ROOT_PATH = SERVER_NAME+':dp/msc-thesis/'
+    REMOTE_LOCAL_PATHS, DOWNLOADS_DIR, BKP_DIR = REMOTE_LOCAL_PATHS_DP, DOWNLOADS_DIR_DP, BKP_DIR_DP
+else:
+    REMOTE_ROOT_PATH = SERVER_NAME+':msc-thesis/'
+
+if opt.quick:
+    REMOTE_LOCAL_PATHS = REMOTE_LOCAL_PATHS[:2]
 
 # delete old folder
 downloads_full_path = os.path.join(os.getcwd(), DOWNLOADS_DIR)
-if os.path.exists(downloads_full_path):
+if os.path.exists(downloads_full_path) and not opt.quick:
     print('Deleting', downloads_full_path)
     shutil.rmtree(downloads_full_path)
